First Complete Build
removed both questions.text and QuestionReader.java because they have become redundant and was having problems having it read the file.

Created the new Question.java, it makes the question list and has two functions that are called by the button click methods in MainActivity.java. One to get a new question and letter and the other to just get a new letter and repeat the question.

Also touched up the GUI
No multi-language support in this version, English only.